Wrote and tested CGSolver
The CGSolver class will provide handle actual running the CG solver on
the fluid sim's linear system, as opposed to mixing all the solver code
up in the class. This should also mean that it's much easier to replace
with a real library if my solver is too slow.

main contains some tests of the solver in two cases: first with a very
simple 4x4 identity matrix system and a second with the 48x48 bcsstk01
matrix and a random b vector.