Making it safe to pass an empty b vector to the CGSolver constructor
This is useful for situations where we want to start uploading everything but don't yet know the b vector we want to solve for, such as when we're creating the solver for the fluid simulation. We'd like to get the matrix and other buffers already uploaded but we don't know the b vector buffer. Also adding the option to pass a cl::Buffer to be used for b instead of a new vector to upload.