Adding kernels to subtract the pressure gradient from velocity
I've got to test these tomorrow but I think they'll be fine. The offsetting
is the reverse of the offsets applied to get the indices in the
velocity_divergence kernel so should be fine. The kernels are seperate
b/c they're run for the velocity grid dimensions instead of the fluid
grid, so x/y velocity runs will have different group dimensions.
I didn't see an obvious painless way to do x/y at once.