Starting to work on advecting values of a buffer
This kernel will be used to advect the velocity fields themselves,
the offsetting is still needed due to the x/y values being offset
from the other fields '0' value. Ie. the first y velocity value is
actually at (0.5, 0) but is treated as being at (0, 0). So we do
still need to offset.