Got bilinear interpolation working properly
Had some issues with the way the element indices were being computed
that caused it to not wrap back around if we were <0 but not <-1. Also
now take into account that blending point 0,0 may be on the other side of
the grid we use a different method for computing the blending point by
simply subtracting off its integer part.

I'm not sure why this integer part subtraction doesn't work for
pos -= convert_int2(pos) or such, but oh well I guess. Next steps are
to actually make use of the blended values in the kernels.