Have written a kernel for applying force to the fluid
It works similar to the set_pixel kernel but I like it a lot more,
I think I'll actually change set_pixel to behave similar to the apply_force
kernel as it makes it trivial to change the brush size, which is nice.

The method works by running the kernel with a global offset to where
the top-left corner of the brush should be placed and with 2d work group
dimensions equal to the w/h of the brush.