Working on initial version of the full sim
I've got everything set up and running in the order discussed in the book
but I seem to have some issues with my solver, keep getting invalid
results back from it. With that missing the sim currently performs
the same behavior as the advect image field but will also
advect the velocity fields. Or at least it should be, I'm not quite
sure if I'm seeing the behavior I expect.