Complete overhaul of the Conjuage Gradient solver
I wrote it in C to get a better understanding of the method and then
split the C program up into kernels/translated to OpenCL. Tests are run
against the identity matrix for a sanity check, a simple example from
wikipedia to solve a problem with a known answer, a single 16x16 fluid grid
and a fluid grid stress test which solves the same system 50 times.
The stress test fluid grid's dimensions are configurable but it's important
to keep in mind the memory usage of the system being solved.

By rewriting this and implementing it properly I've solved
the issues that plagued my likely horribly incorrect solver initially.
Solutions are consistent (checked in the stress test) and fast,
at least on the fluid grids and even for large fluid grids
the solver is pretty quick (<8ms on my desktop).

Next I'll have to make use of the new solver inside the fluid simulation
and see if it all works!